Monolithic commit (sorry)
 - Bugfix: keeps commits in the order they need to be done in, rather
than date order
 - Enhancement: uses newer JGit
 - Enhancement: uses RTC 6.0.6
 - Enhancement: batch mode to allow a migration to be done in stages
(used to defend against a "nothing" outcome from "all or nothing"
migrations)
 - Enhancement: update mode to allow a migration to continue where we
left off (allowing you to keep a git repo "in sync" with an RTC stream
for ease of migration)
 - Bugfix: lots of defensive error-handling code for when RTC fails to
get the changes it was told to obtain
 - ... or fails to apply those changes to the local filesystem correctly
 - Enhancement: can now set commit-comment syntax to link back to RTC
 - Enhancement: can now set git-notes just as easily as comments (use
with caution if using git-lfs)
 - Bugfix: ensure git tags only use legal characters
 - Bugfix: improved handling of .gitignore files to avoid confusing RTC
 - Enhancement: can ignore our own ignore files in case we need to
commit "ignored" files
 - Enhancement: can preserve empty folders if desired
 - Enhancement: much more verbose logging in order to facilitate
diagnosing discrepancies
 - Bugfix: improve error handling
 - Enhancement: initial commit date is now configurable (e.g. to 00:00
on 01/01/2000)
 - Enhancement: autocrlf is now configurable
 - Bugfix: stream removing a component no longer NPEs
 - Bugfix: supports --include-root flag to cope with streams whose
components contain files/paths with the same name
 - Enhancement: git commits now know both the committer and the author
(from RTC's "added to stream" and "author").
 - Enhancement: supports preemptive changeset gap filling mode (but use
with caution as RTC is prone to lying)
 - Bugfix: ensure we don't leak file resources
 - Multiple useful shell scripts to aid migration

Note: Now needs to run using Java8 (due to JGit) not the JVM supplied with RTC6. The "make a Dockerfile" code will arrange for this to happen but if you're not using that, you'll need to remove the built-in JVM yourself.